requirements.txt
kaleido==0.2.1
ffmpeg-python==0.2.0
matplotlib==3.9.0
opencv-python==4.10.0.84
pims==0.7
plotly==5.22.0
seaborn==0.13.2
supervision==0.22.0
ultralytics==8.2.0
streamlit==1.37.0
moviepy==1.0.3
parse==1.20.2
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
tqdm==4.66.4
python-dotenv==1.0.1
requests==2.32.3
pytest==8.2.0
anthropic>=0.28.0
scipy>=1.13.0
